Refactor and separate JS translation parsing
This is in preparation to re-used the code in gnerating translated
scripts.
